Dev1.0

+++ moyu-common/moyu-common-core/src/main/java/cn/xueden/common/core/utils/LayerData.java
@@ -12,7 +12,7 @@ import java.util.Set;
 public class LayerData<T> {
     private Integer code = 0;
 
-    private Integer count;
+    private Long count;
 
     private List<T> data;
 
@@ -28,11 +28,11 @@ public class LayerData<T> {
         this.code = code;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 
-    public void setCount(Integer count) {
+    public void setCount(Long count) {
         this.count = count;
     }
